+++ zrtp/crypto/twoCFB.h
@@ -22,6 +22,7 @@
 #define _TWOCFB_H__
 
 #include <cstdint>
+#include <cstddef>
 
 /**
  * @file aesCFB.h
@@ -55,7 +56,7 @@
  *    Length of the data in bytes
  */
 
-void twoCfbEncrypt(uint8_t* key, int32_t keyLength, uint8_t* IV, uint8_t *data, int32_t dataLength);
+void twoCfbEncrypt(uint8_t* key, size_t keyLength, uint8_t* IV, uint8_t *data, size_t dataLength);
 
 /**
  * Decrypt data with Twofish CFB mode, full block feedback size.
@@ -77,7 +78,7 @@ void twoCfbEncrypt(uint8_t* key, int32_t keyLength, uint8_t* IV, uint8_t *data, 
  *    Length of the data in bytes
  */
 
-void twoCfbDecrypt(uint8_t* key, int32_t keyLength, uint8_t* IV, uint8_t *data, int32_t dataLength);
+void twoCfbDecrypt(uint8_t* key, size_t keyLength, uint8_t* IV, uint8_t *data, size_t dataLength);
 /**
  * @}
  */
